feat: moved popup and added to both create & edit
